// src/types/portfolio.ts
export interface SkillItem {
  n: string; // name
  i: string; // icon tag
  c: string; // background color
  t: string; // text color
}

export interface ProjectItem {
  name: string;
  meta: string;
  desc: string;
  url: string;
}

export interface TimelineItem {
  title: string;
  meta: string;
}